First attempt to implement universal chain: combinator.
In theory it should be possible to create a promise chain by
giving chain: a list of blocks of various types. These blocks
have to be of the same type of one of the blocks given to
fulfilled:, failed:, progressed:, then: or rescue:

We still miss new tests to verify the new behavior and proper
documentation.